processmetrics: explicit type conversion to support arm64

+++ pkg/private/processmetrics/processmetrics_linux.go
@@ -123,7 +123,7 @@ func (c *procStatCollector) updateStat() error {
 	if err != nil {
 		return err
 	}
-	newCount := taskStat.Nlink - 2
+	newCount := uint64(taskStat.Nlink - 2)
 	if newCount != c.lastTaskCount {
 		c.taskListUpdates++
 		c.myProcs, err = procfs.AllThreads(c.myPid)
